Create project and geography store modules
