+++ assets/spa/App.vue
@@ -14,7 +14,7 @@
     </div>
     <footer class="footer row align-items-center bg-info mx-0">
       <div class="col-12 mx-auto text-center text-white">
-        &copy; 2019 International Livestock Reasearch Institute
+        &copy; 2019 International Livestock Research Institute
       </div>
     </footer>
   </div>
@@ -39,7 +39,7 @@
    *    > n = nodes
    */
 
-import NavBar from './components/NavBar.vue';
+import NavBar from './components/NavBar';
 
 export default {
   name: 'App',
